+++ project/auth-client/src/routes/Login.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export default function Login() {
+export default function LoginView() {
   return (
     <div>
       this is the login page
